FAST-API-TH-main/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3

# Google Gemini AI
google-generativeai==0.3.2

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.2.0
matplotlib==3.8.2

# Others
python-dotenv==1.0.0
